First commit for PHP 5 -> 7 conversion
Note: This is not backwards compatbile with PHP5. Considering the
low volume of changes to the pecl lib, this seemed a clean time to
do this break with future changes able to be back ported. Happy to
revisit this decision!

This first commit doesn't remove all of the PHP5 conditional blocks
(many ifdefs exist still, TSRMLS_* still linger, etc.). But I've
explicitly not tried to branch. There's also a little bit of
whitespace clean up here and there.

Most of the changes are:
    - subbing php_gmagick_object, php_gmagickdraw_object, php_gmagickpixel_object
    - zend_object_value -> zend_object * in those functions' returns
    - removing third param from ZVAL_STRING and similar methods
    - switching out hash table loop, removing temp vars from inside loop
    - some int -> size_t subs for param lengths

Not all tests are passing yet (~95% PASS) but the code should compile