Added cleaner hack for queryFontMetrics.
There was an existing hack in queryFontMetrics to allow it to be called when the image list is empty. That hack depended on adding an image temporarily and then removing it through MagickRemoveImage. Unfortunately adding and removing images in GraphicsMagick appears to be fundamentally broken. The code change is to use a completely separate image as a new and improved hack to allow queryFontMetrics to work on empty image lists.